+++ radioactive/utilities.py
@@ -16,8 +16,8 @@ from rich.table import Table
 from rich.text import Text
 from zenlog import log
 
+from radioactive.ffplay import kill_background_ffplays
 from radioactive.last_station import Last_station
-from radioactive.player import kill_background_ffplays
 from radioactive.recorder import record_audio_auto_codec, record_audio_from_url
 
 RED_COLOR = "\033[91m"
@@ -96,7 +96,9 @@ def handle_record(
 
     elif not record_file_path:
         log.debug("filepath: fallback to default path")
-        record_file_path = os.path.join(os.path.expanduser("~"), "Music/radioactive")
+        record_file_path = os.path.join(
+            os.path.expanduser("~"), "Music/radioactive"
+        )  # fallback path
         try:
             os.makedirs(record_file_path, exist_ok=True)
         except Exception as e:
@@ -465,7 +467,7 @@ def handle_user_choice_from_search_result(handler, response):
             print()
             sys.exit(0)
 
-        if user_input == ("y" or "Y"):
+        if user_input in ["y", "Y"]:
             log.debug("Playing UUID from single response")
             global_current_station_info = response[0]
 
@@ -491,6 +493,9 @@ def handle_user_choice_from_search_result(handler, response):
                 # pick a random integer withing range
                 user_input = randint(1, len(response) - 1)
                 log.debug(f"Radom station id: {user_input}")
+            # elif user_input in ["f", "F", "fuzzy"]:
+            # fuzzy find all the stations, and return the selected station id
+            # user_input = fuzzy_find(response)
 
             user_input = int(user_input) - 1  # because ID starts from 1
             if user_input in range(0, len(response)):
@@ -599,3 +604,12 @@ def handle_station_name_from_headers(url):
             )
         )
     return station_name
+
+
+def handle_play_random_station(alias):
+    """Select a random station from favorite menu"""
+    log.debug("playing a random station")
+    alias_map = alias.alias_map
+    index = randint(0, len(alias_map) - 1)
+    station = alias_map[index]
+    return station["name"], station["uuid_or_url"]